# Make dict repr more efficient (50% faster!)