--- a/vm/src/obj/objset.rs
+++ b/vm/src/obj/objset.rs
@@ -234,7 +234,8 @@ impl PySetInner {
     }
 
     fn pop(&self, vm: &VirtualMachine) -> PyResult {
-        if let Some((key, _)) = self.content.pop_front() {
+        // TODO: should be pop_front, but that requires rearranging every index
+        if let Some((key, _)) = self.content.pop_back() {
             Ok(key)
         } else {
             let err_msg = vm.ctx.new_str("pop from an empty set");