```
❯ hyperfine -L X norm,opt './rustpython-{X} pystone.py 50000'    
Benchmark #1: ./rustpython-norm pystone.py 50000
  Time (mean ± σ):     14.349 s ±  0.561 s    [User: 14.295 s, System: 0.010 s]
  Range (min … max):   13.580 s … 14.982 s    10 runs
 
Benchmark #2: ./rustpython-opt pystone.py 50000
  Time (mean ± σ):     9.385 s ±  0.193 s    [User: 9.308 s, System: 0.016 s]
  Range (min … max):   8.986 s … 9.623 s    10 runs
 
Summary
  './rustpython-opt pystone.py 50000' ran
    1.53 ± 0.06 times faster than './rustpython-norm pystone.py 50000'
```
Based on this python version of the data structure (ran through lib2to3): https://code.activestate.com/recipes/578375/

I chose pystone to benchmark this because there aren't many stdlib or even builtins calls, it's just loops, `Load`/`StoreName`s, and arithmetic, so this is just the speedups for module/scope lookups, I think.