@youknowone 2020-09-25
it seems like we are reinventing hashmap