@youknowone 2020-09-25
I think we are using plain hashmap for `__dict__`. how improving `dict` can be related to module/scope lookup?