@coolreader18 2020-09-25
Which `__dict__`? For classes we're using a `HashMap<String, PyObjectRef>`, but in `crate::scope::Scope` it's a PyDictRef for globals and locals, and modules use `PyObject.dict: PyDictRef` as well