@coolreader18 2020-09-25
And `obj.__dict__` is a `dict` as well, of course