@coolreader18 2020-09-25
> it seems like we are reinventing hashmap

Yeah, sort of. I looked into repurposing hashbrown, but it's fundamentally designed for a different kind of use than python dicts; it expects all hash operations to be infallible or it may drop entries, the simd stuff makes it more complex, and it's got a lot of unsafe code that the simpler implementation doesn't, which makes it harder to hack on since you always have to be worrying about safety.