@youknowone 2020-09-25
I got the other points but what about hash infallible? What kind of case be a problem in hashbrown but not in python?