@coolreader18 2020-09-25
> I got the other points but what about hash infallible? What kind of case be a problem in hashbrown but not in python?

Specifically [here](https://docs.rs/hashbrown/0.9.0/src/hashbrown/raw/mod.rs.html#787-802) in hashbrown; when it rearranges/compacts the storage, it needs to know their hashes and if it can't get one that it sorta throws its hands up and deletes what it hasn't touched yet (I think, again, the code is complex). There's a few other places where it does that as well, you can grep for a call to `guard`