@youknowone 2020-09-25
I thought we would replace only hashmap to RawTable if we use hashbrown same as our previous design just like this PR re placing it into a vec. In that case only ints are placed in hashbrown bucket and then the values will be always hashable.