@coolreader18 2020-09-25
Even with RawTable, there are still a few functions that take a hasher `Fn(&T) -> u64`, and those are the functions that are expected to be infallible; even if they were `Fn(&T) -> Result<u64, E>` there's no way to recover from an `Err` with the layout/strategy/whatever hashbrown uses