@youknowone 2020-09-25
I still don't get what's the problem there. when the bucket of RawTable is `(isize, isize)` T in hasher will be also `isize`. It looks like infallible. Will you tell me more details of your design when you thought about hashbrown?