@coolreader18 2020-09-25
Wait, you're saying just to make `indices` a RawTable? Cause a `RawTable<(isize, isize)>` shouldn't be any different than `HashMap<(isize, isize)>`; that's what HashMap is inside:
```rust
pub struct HashMap<K, V, S = DefaultHashBuilder> {
    pub(crate) hash_builder: S,
    pub(crate) table: RawTable<(K, V)>,
}
```