@coolreader18 2020-09-25
And RawTable still has a vector of elements inside, but if we already know where to find the indices for a hash with perturb and stuff then it's more efficient to just have the `Vec<EntryIndex>` than all the machinery of a hashmap