@coolreader18 2020-09-25
Cause barring hash collisions (or masked hash collisions?) it will basically be O(1), just arithmetic to the hashvalue and then index into `indices`, whereas hashbrown's gotta do a whole separate lookup which is still O(1), but just extra/unnecessary stuff when we could do it with just the index