@coolreader18 2020-09-25
Idk, I'm not trying to be aggressive about this or anything, I'm just sorta working out why this works better for myself as well cause it does seem like a HashMap might be more efficient than a sparse array, but it turns out it's not