@youknowone 2020-09-26
Thank you for explanation.

> Idk, I'm not trying to be aggressive about this or anything, I'm just sorta working out why this works better for myself as well cause it does seem like a HashMap might be more efficient than a sparse array, but it turns out it's not

For this, my view was the opposite. Because RawTable is just a wrapper for dynamic array, there will be not much cost comparing to custom vec. If we use custom hasher wiithout actual hashing, I expected the cost wiil be very similar to the custom vec. But simd optimization would be the free lunch in this case. I wonder why there is the difference.